requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
redis==5.0.8
sqlparse==0.5.1
certifi==2024.7.4
cffi @ file:///private/tmp/cffi-20231211-40239-j25k6o/cffi-1.16.0
cryptography==42.0.8
pycparser==2.22
PyQt3D==5.15.6
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
setuptools @ file:///private/tmp/python-setuptools-20240223-4585-1z5qt4/setuptools-69.1.1
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.4/libexec/wheel-0.43.0-py3-none-any.whl#sha256=ff23205a590f2b902a2a21151019e96b93ab4a1409d7c9748b6bd210f4c0f44d

